# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sereni API - Emotional support platform
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  ageRange?: string | null;
  /** @nullable */
  country?: string | null;
  languageCode: string;
  /** @nullable */
  theme?: string | null;
  createdAt: Date;
}